refactor:rename-skinttest->skintest(#79)
