+++ src/main/java/com/swyp3/skin/api/v1/skintest/controller/SkinTestController.java
@@ -5,16 +5,15 @@ import com.swyp3.skin.api.v1.skintest.dto.request.SkinTestPreviewRequest;
 import com.swyp3.skin.api.v1.skintest.dto.response.*;
 import com.swyp3.skin.api.v1.skintest.mapper.SkinInputMapper;
 import com.swyp3.skin.api.v1.skintest.mapper.SkinTestPreviewResponseMapper;
-import com.swyp3.skin.domain.skinttest.dto.SkinPreviewCacheValue;
-import com.swyp3.skin.domain.skinttest.service.SkinPreviewCacheService;
-import com.swyp3.skin.domain.skinttest.service.SkinTestApplicationService;
+import com.swyp3.skin.domain.skintest.dto.SkinPreviewCacheValue;
+import com.swyp3.skin.domain.skintest.service.SkinPreviewCacheService;
+import com.swyp3.skin.domain.skintest.service.SkinTestApplicationService;
 import com.swyp3.skin.global.auth.CustomUserDetails;
 import com.swyp3.skin.global.response.dto.ApiResponse;
 import com.swyp3.skin.recommendation.model.RecommendationResult;
 import com.swyp3.skin.recommendation.model.SkinInput;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Min;
 import lombok.RequiredArgsConstructor;
