+++ src/main/java/com/swyp3/skin/domain/skintest/exception/SkinTestException.java
@@ -1,4 +1,4 @@
-package com.swyp3.skin.domain.skinttest.exception;
+package com.swyp3.skin.domain.skintest.exception;
 
 import com.swyp3.skin.global.exception.CustomException;
 import com.swyp3.skin.global.exception.ErrorCode;
